requirements.txt
furl
yt-dlp
ffmpeg
selenium-wire